Move cloud tasks utils to utils folder
