+++ lib/dlx.js
@@ -2,7 +2,7 @@ import config from "exp-config";
 import { PubSub } from "@google-cloud/pubsub";
 import { logger } from "lu-logger";
 
-import { filterUndefinedNullValues } from "./utils.js";
+import { filterUndefinedNullValues } from "./utils/sequences.js";
 
 const pubSubClient = new PubSub();
