+++ lib/middleware/debug-meta.js
@@ -1,6 +1,6 @@
 import camelCase from "camelcase";
 
-import { filterUndefinedNullValues } from "../utils.js";
+import { filterUndefinedNullValues } from "../utils/sequences.js";
 
 export function setDebugMeta(req, res, next) {
   const debugPrefix = "x-debug-meta";
