+++ lib/publish-task.js
@@ -5,9 +5,9 @@ import gax from "google-gax";
 import * as uuid from "uuid";
 import crypto from "crypto";
 
-import { filterUndefinedNullValues } from "./utils.js";
-import setTimer from "../utils/timer.js";
-import withRetries from "../utils/retry.js";
+import { filterUndefinedNullValues } from "./utils/sequences.js";
+import setTimer from "./utils/timer.js";
+import withRetries from "./utils/retry.js";
 
 const { queues, selfUrl, localPort, useHttpFallback } = config.cloudTasks || {};
 const cloudTasksClient = getCloudTasksClient();
