+++ lib/router.js
@@ -5,7 +5,7 @@ import * as uuid from "uuid";
 import { router as middlewareRouter } from "./middleware/index.js";
 import { errorHandler } from "./error-handler.js";
 import { publishTask, publishTasksBulk } from "./publish-task.js";
-import { appendData, buildNextKeyMapper, buildUrl, keyToUrl, sequenceIterator } from "./utils.js";
+import { appendData, buildNextKeyMapper, buildUrl, keyToUrl, sequenceIterator } from "./utils/sequences.js";
 import buildContext from "./context.js";
 import jobStorage from "./job-storage/index.js";
 import { sendToDlx } from "./dlx.js";
