+++ lib/utils/sequences.js
